refactor(prjx): shared paginator, validate totalCount, fail closed
Address review: extract the duplicated pool pagination into helpers/prjx.ts,
fetch the remaining pages concurrently via PromisePool, validate totalCount is
a non-negative safe integer, and throw if fewer pools than totalCount come back
rather than under-reporting. Both adapters use the shared helper.
